refactor(shared): extract execute_direct_tool_call into shared.tool_execution
Direct tool execution should follow one event-call shape across plugin bundles. Move the helper beside the rest of the shared tool execution code.
