+++ src/owui_ext/tools/multi_model_council.py
@@ -32,6 +32,7 @@ from owui_ext.shared.prompt_utils import (
 from owui_ext.shared.terminal_events import emit_terminal_tool_event
 from owui_ext.shared.tool_event_metadata import CITATION_TOOLS, TERMINAL_EVENT_TOOLS
 from owui_ext.shared.tool_execution import (
+    execute_direct_tool_call,
     normalize_terminal_tools_result,
     process_tool_result,
 )
@@ -112,34 +113,6 @@ def coerce_user_valves(raw_valves: Any, valves_cls: Type[BaseModel]) -> BaseMode
     else:
         data = {}
     return valves_cls.model_validate(data)
-
-
-async def execute_direct_tool_call(
-    *,
-    tool_function_name: str,
-    tool_function_params: dict,
-    tool: dict,
-    extra_params: dict,
-) -> Any:
-    """Execute direct tools through __event_call__ like core middleware."""
-    event_call = extra_params.get("__event_call__")
-    if not callable(event_call):
-        raise RuntimeError("Direct tool execution requires __event_call__ context")
-
-    metadata = extra_params.get("__metadata__")
-    session_id = metadata.get("session_id") if isinstance(metadata, dict) else None
-    return await event_call(
-        {
-            "type": "execute:tool",
-            "data": {
-                "id": str(uuid.uuid4()),
-                "name": tool_function_name,
-                "params": tool_function_params,
-                "server": tool.get("server", {}),
-                "session_id": session_id,
-            },
-        }
-    )
 
 
 def parse_model_ids(value: Any) -> List[str]:
