+++ src/owui_ext/tools/parallel_tools.py
@@ -30,6 +30,7 @@ from owui_ext.shared.async_utils import maybe_await
 from owui_ext.shared.terminal_events import emit_terminal_tool_event
 from owui_ext.shared.tool_event_metadata import CITATION_TOOLS, TERMINAL_EVENT_TOOLS
 from owui_ext.shared.tool_execution import (
+    execute_direct_tool_call,
     normalize_terminal_tools_result,
     process_tool_result,
 )
@@ -101,34 +102,6 @@ async def resolve_terminal_id_from_request_and_metadata(
     return metadata_terminal_id
 
 
-async def execute_direct_tool_call(
-    *,
-    tool_name: str,
-    tool_args: dict,
-    tool: dict,
-    extra_params: dict,
-) -> Any:
-    """Execute direct tools through __event_call__ like core middleware."""
-    event_call = extra_params.get("__event_call__")
-    if not callable(event_call):
-        raise RuntimeError("Direct tool execution requires __event_call__ context")
-
-    metadata = extra_params.get("__metadata__")
-    session_id = metadata.get("session_id") if isinstance(metadata, dict) else None
-    return await event_call(
-        {
-            "type": "execute:tool",
-            "data": {
-                "id": str(uuid.uuid4()),
-                "name": tool_name,
-                "params": tool_args,
-                "server": tool.get("server", {}),
-                "session_id": session_id,
-            },
-        }
-    )
-
-
 async def execute_single_tool(
     tool_name: str,
     tool_args: dict,
@@ -171,8 +144,8 @@ async def execute_single_tool(
 
         if direct_tool:
             result = await execute_direct_tool_call(
-                tool_name=tool_name,
-                tool_args=filtered_args,
+                tool_function_name=tool_name,
+                tool_function_params=filtered_args,
                 tool=tool,
                 extra_params=extra_params,
             )
