+++ tools/parallel_tools.py
@@ -119,6 +119,7 @@ TERMINAL_EVENT_TOOLS: set[str] = {
 
 # --- inlined from src/owui_ext/shared/tool_execution.py (owui_ext.shared.tool_execution) ---
 import json
+import uuid
 from typing import Any, Optional
 from fastapi import Request
 _core_process_tool_result = None
@@ -188,6 +189,33 @@ async def process_tool_result(
     elif tool_result is not None and not isinstance(tool_result, str):
         tool_result = str(tool_result)
     return tool_result, [], []
+
+
+async def execute_direct_tool_call(
+    *,
+    tool_function_name: str,
+    tool_function_params: dict,
+    tool: dict,
+    extra_params: dict,
+) -> Any:
+    """Execute direct tools through ``__event_call__`` like core middleware."""
+    event_call = extra_params.get("__event_call__")
+    if not callable(event_call):
+        raise RuntimeError("Direct tool execution requires __event_call__ context")
+    metadata = extra_params.get("__metadata__")
+    session_id = metadata.get("session_id") if isinstance(metadata, dict) else None
+    return await event_call(
+        {
+            "type": "execute:tool",
+            "data": {
+                "id": str(uuid.uuid4()),
+                "name": tool_function_name,
+                "params": tool_function_params,
+                "server": tool.get("server", {}),
+                "session_id": session_id,
+            },
+        }
+    )
 
 
 def normalize_terminal_tools_result(
@@ -387,34 +415,6 @@ async def resolve_terminal_id_from_request_and_metadata(
     return metadata_terminal_id
 
 
-async def execute_direct_tool_call(
-    *,
-    tool_name: str,
-    tool_args: dict,
-    tool: dict,
-    extra_params: dict,
-) -> Any:
-    """Execute direct tools through __event_call__ like core middleware."""
-    event_call = extra_params.get("__event_call__")
-    if not callable(event_call):
-        raise RuntimeError("Direct tool execution requires __event_call__ context")
-
-    metadata = extra_params.get("__metadata__")
-    session_id = metadata.get("session_id") if isinstance(metadata, dict) else None
-    return await event_call(
-        {
-            "type": "execute:tool",
-            "data": {
-                "id": str(uuid.uuid4()),
-                "name": tool_name,
-                "params": tool_args,
-                "server": tool.get("server", {}),
-                "session_id": session_id,
-            },
-        }
-    )
-
-
 async def execute_single_tool(
     tool_name: str,
     tool_args: dict,
@@ -457,8 +457,8 @@ async def execute_single_tool(
 
         if direct_tool:
             result = await execute_direct_tool_call(
-                tool_name=tool_name,
-                tool_args=filtered_args,
+                tool_function_name=tool_name,
+                tool_function_params=filtered_args,
                 tool=tool,
                 extra_params=extra_params,
             )
